// edelphi/src/main/typescript/src/generated/client/models/Panel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eDelphi REST API
 * REST API for eDelphi
 *
 * The version of the OpenAPI document: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PanelAccessLevel,
    PanelAccessLevelFromJSON,
    PanelAccessLevelFromJSONTyped,
    PanelAccessLevelToJSON,
    PanelState,
    PanelStateFromJSON,
    PanelStateFromJSONTyped,
    PanelStateToJSON,
} from './';

/**
 * 
 * @export
 * @interface Panel
 */
export interface Panel {
    /**
     * 
     * @type {number}
     * @memberof Panel
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof Panel
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Panel
     */
    readonly urlName?: string;
    /**
     * 
     * @type {string}
     * @memberof Panel
     */
    description?: string;
    /**
     * 
     * @type {PanelAccessLevel}
     * @memberof Panel
     */
    accessLevel: PanelAccessLevel;
    /**
     * 
     * @type {PanelState}
     * @memberof Panel
     */
    state: PanelState;
}

export function PanelFromJSON(json: any): Panel {
    return PanelFromJSONTyped(json, false);
}

export function PanelFromJSONTyped(json: any, ignoreDiscriminator: boolean): Panel {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': json['name'],
        'urlName': !exists(json, 'urlName') ? undefined : json['urlName'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'accessLevel': PanelAccessLevelFromJSON(json['accessLevel']),
        'state': PanelStateFromJSON(json['state']),
    };
}

export function PanelToJSON(value?: Panel | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'description': value.description,
        'accessLevel': PanelAccessLevelToJSON(value.accessLevel),
        'state': PanelStateToJSON(value.state),
    };
}